chore: remove old implementation
