+++ src/query/storages/system/src/lib.rs
@@ -60,7 +60,7 @@ mod table_functions_table;
 mod tables_table;
 mod task_history_table;
 mod tasks_table;
-mod temp_files_table_new;
+mod temp_files_table;
 mod temporary_tables_table;
 mod user_functions_table;
 mod users_table;
@@ -125,7 +125,7 @@ pub use task_history_table::parse_task_runs_to_datablock;
 pub use task_history_table::TaskHistoryTable;
 pub use tasks_table::parse_tasks_to_datablock;
 pub use tasks_table::TasksTable;
-pub use temp_files_table_new::TempFilesTable;
+pub use temp_files_table::TempFilesTable;
 pub use temporary_tables_table::TemporaryTablesTable;
 pub use user_functions_table::UserFunctionsTable;
 pub use users_table::UsersTable;
